requirements.txt
discord.py
python-dotenv
psycopg2
urllib3
onnxruntime
Pillow
pgvector
